requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.2
attrs==25.4.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.2
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.12.0
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-jazzmin==3.0.1
django-unfold==0.67.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
frozenlist==1.8.0
idna==3.10
inflection==0.5.1
kombu==5.5.4
multidict==6.7.0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.4.0
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
yarl==1.22.0
